Refactor manager
